Split hub.rs into hub/ module for better organization
hub.rs was 1605 lines. Now split into:
- hub/buffer.rs (174 lines) - HubBuffer trait, BufferStrategy
- hub/mod.rs (977 lines) - Hub builder, MessageSender, tests
- hub/runner.rs (455 lines) - HubRunner, flush_loop, partition

No functional changes. All 18 hub tests pass.
