+++ gateway/src/hub/mod.rs
@@ -32,191 +32,20 @@
 //!     .build();
 //! ```
 
-use crate::buffer_lockfree::LockFreeBuffer;
-use crate::buffer_tiered::TieredBuffer;
+mod buffer;
+mod runner;
+
+pub use buffer::{BufferStrategy, HubBuffer};
+pub use runner::HubRunner;
+
 use crate::checkpoint::CheckpointStore;
-use crate::emit::{Emitter, Event};
+use crate::emit::Emitter;
 use crate::error::PluginError;
 use crate::message::Message;
-use crate::metrics::Metrics;
 use crate::middleware::{Middleware, MiddlewareChain};
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::AtomicU64;
 use tokio::sync::mpsc;
-use tracing::{debug, error, info, warn};
-
-// ============================================================================
-// Buffer abstraction for pluggable buffer strategies
-// ============================================================================
-
-/// Trait for buffer implementations used by Hub
-///
-/// This allows swapping between different buffer strategies:
-/// - `LockFreeBuffer`: Fast, simple, drops on overflow
-/// - `TieredBuffer`: Compressed overflow for graceful degradation
-pub trait HubBuffer: Send + Sync {
-    /// Push a message into the buffer
-    ///
-    /// Returns `true` if stored, `false` if dropped.
-    fn push(&self, msg: Message) -> bool;
-
-    /// Drain up to `n` messages from the buffer
-    fn drain(&self, n: usize) -> Vec<Message>;
-
-    /// Current number of messages in the buffer
-    fn len(&self) -> usize;
-
-    /// Check if buffer is empty
-    fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Buffer capacity
-    fn capacity(&self) -> usize;
-
-    /// Strategy name for logging
-    fn strategy_name(&self) -> &'static str;
-}
-
-impl HubBuffer for LockFreeBuffer {
-    fn push(&self, msg: Message) -> bool {
-        LockFreeBuffer::push(self, msg)
-    }
-
-    fn drain(&self, n: usize) -> Vec<Message> {
-        LockFreeBuffer::drain(self, n)
-    }
-
-    fn len(&self) -> usize {
-        LockFreeBuffer::len(self)
-    }
-
-    fn capacity(&self) -> usize {
-        LockFreeBuffer::capacity(self)
-    }
-
-    fn strategy_name(&self) -> &'static str {
-        "standard"
-    }
-}
-
-impl HubBuffer for TieredBuffer {
-    fn push(&self, msg: Message) -> bool {
-        TieredBuffer::push(self, msg)
-    }
-
-    fn drain(&self, n: usize) -> Vec<Message> {
-        TieredBuffer::drain(self, n)
-    }
-
-    fn len(&self) -> usize {
-        TieredBuffer::len(self)
-    }
-
-    fn capacity(&self) -> usize {
-        TieredBuffer::capacity(self)
-    }
-
-    fn strategy_name(&self) -> &'static str {
-        "tiered"
-    }
-}
-
-/// Buffer strategy configuration for Hub
-///
-/// Controls how the Hub handles message buffering and overflow.
-#[derive(Clone)]
-pub enum BufferStrategy {
-    /// Standard lock-free buffer (default)
-    ///
-    /// Fast and simple. When full, new messages are dropped.
-    /// Best for: steady-state traffic, low-latency requirements.
-    Standard {
-        /// Buffer capacity in messages
-        capacity: usize,
-    },
-
-    /// Tiered buffer with compressed overflow
-    ///
-    /// Primary buffer (fast) + secondary buffer (compressed).
-    /// When primary fills, messages overflow to compressed secondary.
-    /// Best for: traffic spikes, graceful degradation.
-    Tiered {
-        /// Primary (hot) buffer capacity
-        primary_capacity: usize,
-        /// Secondary (compressed overflow) buffer capacity
-        secondary_capacity: usize,
-        /// Compression level (1-22, default 3)
-        compression_level: i32,
-    },
-}
-
-impl BufferStrategy {
-    /// Create a standard buffer strategy
-    pub fn standard(capacity: usize) -> Self {
-        Self::Standard { capacity }
-    }
-
-    /// Create a tiered buffer strategy with default compression
-    pub fn tiered(primary_capacity: usize, secondary_capacity: usize) -> Self {
-        Self::Tiered {
-            primary_capacity,
-            secondary_capacity,
-            compression_level: 3,
-        }
-    }
-
-    /// Create a tiered buffer with custom compression level
-    pub fn tiered_with_compression(
-        primary_capacity: usize,
-        secondary_capacity: usize,
-        compression_level: i32,
-    ) -> Self {
-        Self::Tiered {
-            primary_capacity,
-            secondary_capacity,
-            compression_level: compression_level.clamp(1, 22),
-        }
-    }
-
-    /// Build the buffer from this strategy
-    fn build(&self) -> Arc<dyn HubBuffer> {
-        match self {
-            Self::Standard { capacity } => Arc::new(LockFreeBuffer::new(*capacity)),
-            Self::Tiered {
-                primary_capacity,
-                secondary_capacity,
-                compression_level,
-            } => Arc::new(
-                TieredBuffer::new(*primary_capacity, *secondary_capacity, 1)
-                    .with_compression_level(*compression_level),
-            ),
-        }
-    }
-
-    /// Get total capacity for this strategy
-    #[allow(dead_code)]
-    fn capacity(&self) -> usize {
-        match self {
-            Self::Standard { capacity } => *capacity,
-            Self::Tiered {
-                primary_capacity,
-                secondary_capacity,
-                ..
-            } => primary_capacity + secondary_capacity,
-        }
-    }
-}
-
-impl Default for BufferStrategy {
-    fn default() -> Self {
-        Self::Standard { capacity: 10_000 }
-    }
-}
-
-// ============================================================================
-// Hub implementation
-// ============================================================================
 
 /// The Hub - central message pipeline
 ///
@@ -435,459 +264,14 @@ impl MessageSender {
     }
 }
 
-/// Hub runner - processes messages through the pipeline
-pub struct HubRunner {
-    rx: mpsc::Receiver<Message>,
-    buffer: Arc<dyn HubBuffer>,
-    batch_size: usize,
-    flush_interval_ms: u64,
-    middleware: MiddlewareChain,
-    emitters: Vec<Arc<dyn Emitter>>,
-    /// Optional checkpoint store for tracking delivery progress
-    checkpoint_store: Option<Arc<dyn CheckpointStore>>,
-    /// Monotonically increasing sequence number for checkpoint tracking
-    sequence: Arc<AtomicU64>,
-}
-
-impl HubRunner {
-    /// Run the hub, processing messages until the channel closes
-    ///
-    /// This will:
-    /// 1. Receive messages from the input channel
-    /// 2. Apply middleware chain
-    /// 3. Buffer messages
-    /// 4. Flush immediately when buffer hits batch_size (inline flush)
-    /// 5. Timer-based flush for partial batches (latency bound)
-    /// 6. On shutdown, drain remaining buffer before exiting
-    pub async fn run(mut self) -> Result<(), PluginError> {
-        info!(
-            emitters = self.emitters.len(),
-            middleware = self.middleware.len(),
-            buffer_capacity = self.buffer.capacity(),
-            buffer_strategy = self.buffer.strategy_name(),
-            "Hub started"
-        );
-
-        // Initialize Prometheus metrics (ignore error if already initialized)
-        if let Ok(metrics) = Metrics::init() {
-            metrics.set_buffer_capacity(self.buffer.capacity());
-        }
-
-        if self.emitters.is_empty() {
-            warn!("No emitters registered - messages will be buffered but not delivered");
-        }
-
-        // Shutdown signal: when true, flush loop will drain and exit
-        let (shutdown_tx, shutdown_rx) = tokio::sync::watch::channel(false);
-
-        // Spawn timer-based flusher for partial batches (latency bound)
-        let buffer = Arc::clone(&self.buffer);
-        let emitters_for_timer = self.emitters.clone();
-        let batch_size = self.batch_size;
-        let flush_interval_ms = self.flush_interval_ms;
-        let checkpoint_store = self.checkpoint_store.clone();
-        let sequence = Arc::clone(&self.sequence);
-        let flush_handle = tokio::spawn(async move {
-            flush_loop(
-                buffer,
-                emitters_for_timer,
-                shutdown_rx,
-                batch_size,
-                flush_interval_ms,
-                checkpoint_store,
-                sequence,
-            )
-            .await;
-        });
-
-        // Process incoming messages
-        while let Some(msg) = self.rx.recv().await {
-            // Record received message
-            if let Some(metrics) = Metrics::get() {
-                metrics.record_received(&msg.source, &msg.message_type, 1);
-            }
-
-            // Apply middleware
-            let processed = self.middleware.process(msg).await;
-
-            if let Some(msg) = processed {
-                debug!(id = %msg.id, "Message buffered");
-                if !self.buffer.push(msg) {
-                    warn!("Buffer overflow, message dropped");
-                    // Record dropped due to buffer overflow
-                    if let Some(metrics) = Metrics::get() {
-                        metrics.record_dropped("buffer_overflow", 1);
-                    }
-                }
-
-                // INLINE FLUSH: When buffer hits threshold, flush immediately
-                // This avoids waiting for the timer when we have a full batch
-                if self.buffer.len() >= self.batch_size {
-                    self.flush_batch().await;
-                }
-            } else {
-                // Message was filtered by middleware
-                if let Some(metrics) = Metrics::get() {
-                    metrics.record_dropped("middleware_filtered", 1);
-                }
-            }
-        }
-
-        // Channel closed - signal flush loop to drain and stop
-        info!("Hub shutting down, draining buffer...");
-        let _ = shutdown_tx.send(true);
-
-        // Wait for flush loop to complete (it will drain the buffer first)
-        if let Err(e) = flush_handle.await {
-            warn!(error = %e, "Flush task failed during shutdown");
-        }
-
-        info!(remaining = self.buffer.len(), "Hub shutdown complete");
-
-        Ok(())
-    }
-
-    /// Flush a batch of messages to emitters
-    ///
-    /// Called inline when buffer hits threshold, and by the timer loop.
-    async fn flush_batch(&self) {
-        let flush_start = std::time::Instant::now();
-        let messages = self.buffer.drain(self.batch_size);
-
-        // Update buffer size metric after drain
-        if let Some(metrics) = Metrics::get() {
-            metrics.set_buffer_size(self.buffer.len());
-        }
-
-        if messages.is_empty() {
-            return;
-        }
-
-        let total_events = messages.len();
-
-        // Assign sequence numbers to this batch for checkpoint tracking.
-        // Each message gets a unique sequence number for precise tracking.
-        let batch_start_seq = self
-            .sequence
-            .fetch_add(messages.len() as u64, Ordering::SeqCst);
-
-        // Convert Messages to proto Events for outputs
-        let events: Vec<Event> = messages
-            .into_iter()
-            .map(Event::from)
-            .collect();
-
-        // Partition events by destination with zero-copy optimization
-        let batches = partition_by_destination_owned(events, &self.emitters);
-
-        // Pre-aggregate metrics to reduce Prometheus HashMap lookups.
-        // Key: (emitter_name, event_type), Value: count
-        // This reduces O(events) lookups to O(unique label combos).
-        let mut forward_counts: std::collections::HashMap<(&str, &str), u64> =
-            std::collections::HashMap::new();
-
-        // Send pre-partitioned batches to each emitter.
-        // Each emitter's checkpoint advances by the number of events IT processed,
-        // not the total batch size (which may differ due to routing).
-        for emitter in &self.emitters {
-            let routed_events = match batches.get(emitter.name()) {
-                Some(events) if !events.is_empty() => events,
-                _ => continue,
-            };
-
-            let emitter_start = std::time::Instant::now();
-            let emit_result = emitter.emit(routed_events).await;
-            let emitter_duration = emitter_start.elapsed();
-
-            // Record per-emitter metrics
-            if let Some(metrics) = Metrics::get() {
-                metrics.record_emitter_flush(
-                    emitter.name(),
-                    routed_events.len(),
-                    emitter_duration,
-                    emit_result.is_ok(),
-                );
-            }
-
-            if let Err(e) = emit_result {
-                error!(
-                    emitter = emitter.name(),
-                    error = %e,
-                    count = routed_events.len(),
-                    "Failed to emit"
-                );
-                if let Some(metrics) = Metrics::get() {
-                    metrics.record_dropped(emitter.name(), routed_events.len() as u64);
-                }
-                // Note: checkpoint NOT updated on failure - emitter is behind
-            } else {
-                debug!(
-                    emitter = emitter.name(),
-                    count = routed_events.len(),
-                    "Emitted (inline)"
-                );
-                // Aggregate counts locally instead of per-event Prometheus updates
-                for event in routed_events {
-                    *forward_counts
-                        .entry((emitter.name(), event.event_type.as_str()))
-                        .or_default() += 1;
-                }
-                // Update checkpoint: advance by number of events this emitter processed.
-                // Uses batch_start_seq + emitter's event count - 1 as the checkpoint.
-                // This ensures each emitter tracks its own progress independently.
-                if let Some(ref store) = self.checkpoint_store {
-                    let emitter_end_seq = batch_start_seq + routed_events.len() as u64 - 1;
-                    store.set(emitter.name(), emitter_end_seq);
-                }
-            }
-        }
-
-        // Batch update to Prometheus - single HashMap lookup per unique label combo
-        if let Some(metrics) = Metrics::get() {
-            metrics.record_forwarded_batch(&forward_counts);
-            metrics.inc_flush();
-
-            // Calculate and record throughput
-            let flush_duration = flush_start.elapsed();
-            if flush_duration.as_secs_f64() > 0.0 {
-                let events_per_sec = total_events as f64 / flush_duration.as_secs_f64();
-                metrics.set_events_per_second(events_per_sec);
-            }
-        }
-    }
-
-    /// Get a reference to the buffer for monitoring
-    pub fn buffer(&self) -> &Arc<dyn HubBuffer> {
-        &self.buffer
-    }
-}
-
-/// Partition events by destination, returning ready-to-emit batches
-///
-/// For events that go to only ONE emitter, moves the event (no clone).
-/// For events that go to MULTIPLE emitters, clones N-1 times (last one is moved).
-///
-/// # Performance
-///
-/// - Single pass to count destinations per event: O(events × emitters)
-/// - Second pass to distribute: O(events × avg_destinations)
-/// - Multi-destination events: N-1 clones instead of N (last destination gets moved value)
-fn partition_by_destination_owned(
-    events: Vec<Event>,
-    emitters: &[Arc<dyn Emitter>],
-) -> std::collections::HashMap<&'static str, Vec<Event>> {
-    if emitters.is_empty() || events.is_empty() {
-        return std::collections::HashMap::new();
-    }
-
-    let mut batches: std::collections::HashMap<&'static str, Vec<Event>> =
-        std::collections::HashMap::new();
-
-    // Pre-allocate for each emitter
-    for emitter in emitters {
-        batches.insert(
-            emitter.name(),
-            Vec::with_capacity(events.len() / emitters.len()),
-        );
-    }
-
-    // First pass: count destinations for each event
-    let mut dest_counts: Vec<usize> = vec![0; events.len()];
-    let mut destinations: Vec<Vec<&'static str>> =
-        vec![Vec::with_capacity(emitters.len()); events.len()];
-
-    for (idx, event) in events.iter().enumerate() {
-        let broadcast = event.route_to.is_empty();
-
-        for emitter in emitters {
-            if broadcast || event.route_to.iter().any(|r| r == emitter.name()) {
-                dest_counts[idx] += 1;
-                destinations[idx].push(emitter.name());
-            }
-        }
-    }
-
-    // Second pass: distribute events
-    // - Single destination: move (no clone)
-    // - Multiple destinations: clone N-1 times, move on final access
-    for (idx, event) in events.into_iter().enumerate() {
-        let count = dest_counts[idx];
-        let dests = &destinations[idx];
-
-        match count {
-            0 => {
-                // Event has no matching emitters, drop it
-            }
-            1 => {
-                // Single destination: move without cloning
-                if let Some(batch) = batches.get_mut(dests[0]) {
-                    batch.push(event);
-                }
-            }
-            _ => {
-                // Multiple destinations: clone for first N-1, move for last
-                // This saves one clone compared to cloning for all N destinations
-                for (i, dest) in dests.iter().enumerate() {
-                    if let Some(batch) = batches.get_mut(*dest) {
-                        if i == dests.len() - 1 {
-                            // Last destination: move the original
-                            batch.push(event);
-                            break;
-                        } else {
-                            // Not last: clone
-                            batch.push(event.clone());
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    batches
-}
-
-/// Background flush loop - sends buffered messages to emitters
-///
-/// When shutdown is signaled, drains the buffer completely before exiting.
-async fn flush_loop(
-    buffer: Arc<dyn HubBuffer>,
-    emitters: Vec<Arc<dyn Emitter>>,
-    mut shutdown_rx: tokio::sync::watch::Receiver<bool>,
-    batch_size: usize,
-    flush_interval_ms: u64,
-    checkpoint_store: Option<Arc<dyn CheckpointStore>>,
-    sequence: Arc<AtomicU64>,
-) {
-    loop {
-        // Check if shutdown requested and buffer is empty
-        if *shutdown_rx.borrow() && buffer.is_empty() {
-            debug!("Flush loop: shutdown complete, buffer drained");
-            break;
-        }
-
-        // Either wait for interval OR shutdown signal
-        tokio::select! {
-            _ = tokio::time::sleep(tokio::time::Duration::from_millis(flush_interval_ms)) => {}
-            _ = shutdown_rx.changed() => {
-                // Shutdown signaled - continue loop to drain buffer
-                debug!("Flush loop: shutdown signaled, draining buffer");
-            }
-        }
-
-        let flush_start = std::time::Instant::now();
-        let messages = buffer.drain(batch_size);
-
-        // Update buffer size metric after drain
-        if let Some(metrics) = Metrics::get() {
-            metrics.set_buffer_size(buffer.len());
-        }
-
-        if messages.is_empty() {
-            continue;
-        }
-
-        let total_events = messages.len();
-
-        // Assign sequence numbers to this batch for checkpoint tracking.
-        // Each message gets a unique sequence number for precise tracking.
-        let batch_start_seq = sequence.fetch_add(messages.len() as u64, Ordering::SeqCst);
-
-        // Convert Messages to proto Events for outputs
-        let events: Vec<Event> = messages
-            .into_iter()
-            .map(Event::from)
-            .collect();
-
-        // Partition events by destination with zero-copy optimization
-        // - Single-destination events: moved, not cloned
-        // - Multi-destination events: N-1 clones (last one moved)
-        let batches = partition_by_destination_owned(events, &emitters);
-
-        // Pre-aggregate metrics to reduce Prometheus HashMap lookups.
-        // Key: (emitter_name, event_type), Value: count
-        // This reduces O(events) lookups to O(unique label combos).
-        let mut forward_counts: std::collections::HashMap<(&str, &str), u64> =
-            std::collections::HashMap::new();
-
-        // Send pre-partitioned batches to each emitter.
-        // Each emitter's checkpoint advances by the number of events IT processed,
-        // not the total batch size (which may differ due to routing).
-        for emitter in &emitters {
-            let routed_events = match batches.get(emitter.name()) {
-                Some(events) if !events.is_empty() => events,
-                _ => continue,
-            };
-
-            let emitter_start = std::time::Instant::now();
-            let emit_result = emitter.emit(routed_events).await;
-            let emitter_duration = emitter_start.elapsed();
-
-            // Record per-emitter metrics
-            if let Some(metrics) = Metrics::get() {
-                metrics.record_emitter_flush(
-                    emitter.name(),
-                    routed_events.len(),
-                    emitter_duration,
-                    emit_result.is_ok(),
-                );
-            }
-
-            if let Err(e) = emit_result {
-                error!(
-                    emitter = emitter.name(),
-                    error = %e,
-                    count = routed_events.len(),
-                    "Failed to emit"
-                );
-                // Record emit failures
-                if let Some(metrics) = Metrics::get() {
-                    metrics.record_dropped(emitter.name(), routed_events.len() as u64);
-                }
-                // Note: checkpoint NOT updated on failure - emitter is behind
-            } else {
-                debug!(
-                    emitter = emitter.name(),
-                    count = routed_events.len(),
-                    "Emitted"
-                );
-                // Aggregate counts locally instead of per-event Prometheus updates
-                for event in routed_events {
-                    *forward_counts
-                        .entry((emitter.name(), event.event_type.as_str()))
-                        .or_default() += 1;
-                }
-                // Update checkpoint: advance by number of events this emitter processed.
-                // Uses batch_start_seq + emitter's event count - 1 as the checkpoint.
-                // This ensures each emitter tracks its own progress independently.
-                if let Some(ref store) = checkpoint_store {
-                    let emitter_end_seq = batch_start_seq + routed_events.len() as u64 - 1;
-                    store.set(emitter.name(), emitter_end_seq);
-                }
-            }
-        }
-
-        // Batch update to Prometheus - single HashMap lookup per unique label combo
-        if let Some(metrics) = Metrics::get() {
-            metrics.record_forwarded_batch(&forward_counts);
-            metrics.inc_flush();
-
-            // Calculate and record throughput
-            let flush_duration = flush_start.elapsed();
-            if flush_duration.as_secs_f64() > 0.0 {
-                let events_per_sec = total_events as f64 / flush_duration.as_secs_f64();
-                metrics.set_events_per_second(events_per_sec);
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 #[allow(clippy::unwrap_used, clippy::expect_used)]
 mod tests {
     use super::*;
-    use crate::emit::StdoutEmitter;
+    use crate::emit::{Event, StdoutEmitter};
     use crate::middleware::{Filter, Transform};
     use bytes::Bytes;
+    use runner::partition_by_destination_owned;
     use std::sync::atomic::{AtomicUsize, Ordering};
 
     #[test]
@@ -981,10 +365,7 @@ mod tests {
             fn name(&self) -> &'static str {
                 "counter"
             }
-            async fn emit(
-                &self,
-                events: &[Event],
-            ) -> Result<(), crate::error::PluginError> {
+            async fn emit(&self, events: &[Event]) -> Result<(), crate::error::PluginError> {
                 self.0.fetch_add(events.len() as u64, Ordering::SeqCst);
                 Ok(())
             }
@@ -1052,10 +433,7 @@ mod tests {
                 "counter"
             }
 
-            async fn emit(
-                &self,
-                events: &[Event],
-            ) -> Result<(), crate::error::PluginError> {
+            async fn emit(&self, events: &[Event]) -> Result<(), crate::error::PluginError> {
                 self.count.fetch_add(events.len() as u64, Ordering::SeqCst);
                 Ok(())
             }
@@ -1429,10 +807,7 @@ mod tests {
                 "timing"
             }
 
-            async fn emit(
-                &self,
-                events: &[Event],
-            ) -> Result<(), crate::error::PluginError> {
+            async fn emit(&self, events: &[Event]) -> Result<(), crate::error::PluginError> {
                 self.count.fetch_add(events.len() as u64, Ordering::SeqCst);
                 let mut first = self.first_emit_time.lock();
                 if first.is_none() {
@@ -1544,10 +919,7 @@ mod tests {
                 "counter"
             }
 
-            async fn emit(
-                &self,
-                events: &[Event],
-            ) -> Result<(), crate::error::PluginError> {
+            async fn emit(&self, events: &[Event]) -> Result<(), crate::error::PluginError> {
                 self.count.fetch_add(events.len() as u64, Ordering::SeqCst);
                 Ok(())
             }
